feat: Apple Silicon向け配布基盤を整備し現行macOS対応を更新
- .github: push workflowからIntel runnerを外し、タグ起点で配布物を公開するrelease workflowを追加
- Assets.xcassets: arm64表記入りのApp Iconソースを追加し、各サイズのアイコンを書き出し直せる構成に更新
- Gas Mask.xcodeproj: 不要な設定を整理し、現行Xcodeで通るビルド設定へ調整
- Source: 非推奨APIや型まわりを見直し、Sparkle未設定時でも安全に動くよう更新
- Swift: 更新確認UIをfork向け運用に合わせ、Sparkle未構成時の表示と制御を追加
- Resources: 自動更新の既定値を無効化し、fork配布方針に合わせて初期設定を調整
- Release: dmg生成スクリプトをdist出力前提に整理し、再利用しやすい引数構成へ変更
- ルート: build.shをarm64既定に見直し、universal用補助スクリプト、package-release.sh、アイコン更新スクリプト、LICENSEを追加
- docs: リリース手順書とリリースノートを追加し、forkの配布フローを文書化
- README.md: fork向けの説明、配布方法、ライセンス、ビルド手順に全面更新

+++ Source/HostsMainController.m
@@ -455,11 +455,11 @@ static HostsMainController *sharedInstance = nil;
 
 - (NSArray*)allHostsFilesGrouped
 {
-	int nrControllers = [controllers count];
+	NSUInteger nrControllers = [controllers count];
 	
 	NSMutableArray *array = [NSMutableArray arrayWithCapacity:nrControllers];
 	
-	for (int i=0; i<nrControllers; i++) {
+	for (NSUInteger i=0; i<nrControllers; i++) {
 		NSObject<HostsControllerProtocol> *controller = [controllers objectAtIndex:i];
 		
 		[array addObject:[Pair pairWithLeft:[controller name] right:[controller hostsFiles]]];
@@ -582,10 +582,10 @@ static HostsMainController *sharedInstance = nil;
 
 - (void)hostsFileRemoved:(Hosts*)hosts controller:(NSObject<HostsControllerProtocol>*)controller
 {
-	int index = [controllers indexOfObject:controller];
+	NSUInteger index = [controllers indexOfObject:controller];
 	NSArray *files = [[[self content] objectAtIndex:index] children];
 	
-	int i;
+	NSUInteger i;
 	for (i=0; i<[files count]; i++) {
 		if ([[files objectAtIndex:i] isEqual:hosts]) {
 			break;
@@ -725,7 +725,7 @@ static HostsMainController *sharedInstance = nil;
 
 - (void)addHostsFile:(Hosts*)hosts forController:(NSObject<HostsControllerProtocol>*)controller
 {
-	int index = [controllers indexOfObject:controller];
+	NSUInteger index = [controllers indexOfObject:controller];
 	
 	NSIndexPath *indexPath = [[NSIndexPath indexPathWithIndex:index] indexPathByAddingIndex:[[controller hostsFiles] count]-1];
 	[self insertObject:hosts atArrangedObjectIndexPath:indexPath];
